is_not_null to module
