+++ polars/polars-lazy/src/physical_plan/expressions/is_not_null.rs
@@ -1,0 +1,32 @@
+use crate::physical_plan::state::ExecutionState;
+use crate::prelude::*;
+use polars_core::prelude::*;
+use std::sync::Arc;
+
+pub struct IsNotNullExpr {
+    physical_expr: Arc<dyn PhysicalExpr>,
+    expr: Expr,
+}
+
+impl IsNotNullExpr {
+    pub fn new(physical_expr: Arc<dyn PhysicalExpr>, expr: Expr) -> Self {
+        Self {
+            physical_expr,
+            expr,
+        }
+    }
+}
+
+impl PhysicalExpr for IsNotNullExpr {
+    fn as_expression(&self) -> &Expr {
+        &self.expr
+    }
+
+    fn evaluate(&self, df: &DataFrame, state: &ExecutionState) -> Result<Series> {
+        let series = self.physical_expr.evaluate(df, state)?;
+        Ok(series.is_not_null().into_series())
+    }
+    fn to_field(&self, _input_schema: &Schema) -> Result<Field> {
+        Ok(Field::new("is_not_null", DataType::Boolean))
+    }
+}
